Update ParentScope example in Composition.md
I rewrote the example with ParentScope to better reflect how ParentScope/DelayParentScope should be used.  The problem with old example is that the local definitions of the parameters do not agree with the parameter list.  If you define equations then the parameters in the equations don't match the parameter list, which leads to a situation where parameters(sys) are different than the parameters in equations(sys)